% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splith.R
\name{splith}
\alias{splith}
\title{Computes split-half internal consistency measures via permutation for human behavior tasks.}
\usage{
splith(
  data,
  outcome = "RT",
  average = "mean",
  permutations = 10,
  variable = NULL,
  condition = NULL,
  subject = "subject",
  include_block = FALSE,
  block = NULL,
  return_iterations = FALSE
)
}
\arguments{
\item{data}{A data frame containing at least an outcome variable, the variable you want to analyze and a variable for subject's identificators.}

\item{outcome}{The name of your outcome variable in strings. The default is "RT".}

\item{average}{The type of averaging that you want to apply for computing split-half.The default is "mean".}

\item{permutations}{Number of times that you want to compute the split-half.Default is 10, but more is recommended.}

\item{variable}{Name of the variable to be analyzed in strings (e.g. "Congruency").Default is set to NULL, and is required to be filled.}

\item{condition}{A variable's name containing different conditions for splitting analysis.This argument is optional.}

\item{subject}{A variable's name for subject's indentification.The default is "subject".}

\item{include_block}{A logical that determine if each half computed should apply constrains depending on blocking. If TRUE, split-half will be computed splitting by block. Default is FALSE.}

\item{block}{The name of the variable that you want to apply constrains in strings. If include_block is TRUE, block is required to be filled. Default is null.}

\item{return_iteration}{A logical controlling if te split-half estimation for each iteration is returned in the output. Default is FALSE.}
}
\value{
A data frame containing split-half and spearman-brown estimators for n permutations.
}
\description{
Computes split-half reliability indexes ussing a permutation approach for the difference in outcome.
}
